requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
cmake==4.0.3
filelock==3.18.0
fsspec==2025.5.1
ftfy==6.3.1
gdown==5.2.0
hf-xet==1.1.4
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
lit==18.1.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open_clip_torch==2.32.0
packaging==25.0
pillow==11.2.1
PySocks==1.7.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
soupsieve==2.7
sympy==1.14.0
timm==1.0.15
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.1
triton==2.0.0
typing_extensions==4.14.0
urllib3==2.5.0
wcwidth==0.2.13
